requirements.txt
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.2.0
beautifulsoup4==4.12.3
bs4==0.0.2
charset-normalizer==3.3.2
Django==5.0.3
frozenlist==1.4.1
idna==3.6
multidict==6.0.5
python-decouple==3.8
requests==2.31.0
soupsieve==2.5
sqlparse==0.4.4
urllib3==2.2.1
yarl==1.9.4
